Create defaults package and move DefaultFunctions into it
