+++ core/src/java/show/ezkz/casprzak/core/functions/unitary/piecewise/Sign.java
@@ -3,7 +3,7 @@ package show.ezkz.casprzak.core.functions.unitary.piecewise;
 import show.ezkz.casprzak.core.functions.GeneralFunction;
 import show.ezkz.casprzak.core.functions.commutative.Product;
 import show.ezkz.casprzak.core.functions.unitary.UnitaryFunction;
-import show.ezkz.casprzak.core.tools.DefaultFunctions;
+import show.ezkz.casprzak.core.tools.defaults.DefaultFunctions;
 
 import java.util.Map;
