+++ core/src/java/show/ezkz/casprzak/core/functions/unitary/trig/normal/Cos.java
@@ -6,7 +6,7 @@ import show.ezkz.casprzak.core.functions.commutative.Product;
 import show.ezkz.casprzak.core.functions.unitary.UnitaryFunction;
 import show.ezkz.casprzak.core.functions.unitary.piecewise.DomainRestrictor;
 import show.ezkz.casprzak.core.functions.unitary.trig.inverse.Acos;
-import show.ezkz.casprzak.core.tools.DefaultFunctions;
+import show.ezkz.casprzak.core.tools.defaults.DefaultFunctions;
 
 import java.util.Map;
