+++ core/src/java/show/ezkz/casprzak/core/functions/unitary/trig/normal/Coth.java
@@ -9,7 +9,7 @@ import show.ezkz.casprzak.core.functions.unitary.piecewise.DomainRestrictor;
 import show.ezkz.casprzak.core.functions.unitary.specialcases.Ln;
 import show.ezkz.casprzak.core.functions.unitary.UnitaryFunction;
 import show.ezkz.casprzak.core.functions.unitary.trig.inverse.Acoth;
-import show.ezkz.casprzak.core.tools.DefaultFunctions;
+import show.ezkz.casprzak.core.tools.defaults.DefaultFunctions;
 
 import java.util.Map;
