+++ core/src/java/show/ezkz/casprzak/core/functions/unitary/trig/normal/Sech.java
@@ -8,7 +8,7 @@ import show.ezkz.casprzak.core.functions.unitary.UnitaryFunction;
 import show.ezkz.casprzak.core.functions.unitary.piecewise.DomainRestrictor;
 import show.ezkz.casprzak.core.functions.unitary.trig.inverse.Asech;
 import show.ezkz.casprzak.core.functions.unitary.trig.inverse.Atan;
-import show.ezkz.casprzak.core.tools.DefaultFunctions;
+import show.ezkz.casprzak.core.tools.defaults.DefaultFunctions;
 
 import java.util.Map;
