+++ core/src/java/show/ezkz/casprzak/core/tools/singlevariable/Solver.java
@@ -6,7 +6,7 @@ import show.ezkz.casprzak.core.functions.binary.Pow;
 import show.ezkz.casprzak.core.functions.commutative.Product;
 import show.ezkz.casprzak.core.functions.endpoint.Constant;
 import show.ezkz.casprzak.core.functions.unitary.piecewise.Abs;
-import show.ezkz.casprzak.core.tools.DefaultFunctions;
+import show.ezkz.casprzak.core.tools.defaults.DefaultFunctions;
 import show.ezkz.casprzak.core.tools.SolverTools;
 import show.ezkz.casprzak.core.tools.VariableTools;
