+++ tests/OutputTest.java
@@ -1,6 +1,6 @@
 import show.ezkz.casprzak.core.functions.endpoint.Constant;
 import org.junit.jupiter.api.Test;
-import show.ezkz.casprzak.core.tools.DefaultFunctions;
+import show.ezkz.casprzak.core.tools.defaults.DefaultFunctions;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
