+++ tests/SubstituteTest.java
@@ -5,7 +5,7 @@ import show.ezkz.casprzak.core.functions.unitary.trig.normal.Cos;
 import show.ezkz.casprzak.core.functions.unitary.trig.normal.Sin;
 import org.junit.jupiter.api.Test;
 import show.ezkz.casprzak.parsing.FunctionParser;
-import show.ezkz.casprzak.core.tools.DefaultFunctions;
+import show.ezkz.casprzak.core.tools.defaults.DefaultFunctions;
 
 import java.util.Map;
